Refactor code to be optimized
